+++ PatternOmatic/nlp/bnf.py
@@ -1,15 +1,20 @@
-""" NLP Engines """
+""" Backus Naur Form Grammar Generator through Spacy NLP Engine """
 from inspect import getmembers
 from spacy.tokens import Doc, Token
 from PatternOmatic.settings.config import Config
-from PatternOmatic.settings.literals import *
+from PatternOmatic.settings.literals import S, P, T, F, OP, NEGATION, ZERO_OR_ONE, ZERO_OR_MORE, ONE_OR_MORE, LENGTH, \
+    XPS, IN, NOT_IN, EQQ, GEQ, LEQ, GTH, LTH, TOKEN_WILDCARD, UNDERSCORE, EF, ORTH, TEXT, LOWER, POS, TAG, DEP, LEMMA, \
+    SHAPE, ENT_TYPE, IS_ALPHA, IS_ASCII, IS_DIGIT, IS_BRACKET, IS_LOWER, IS_PUNCT, IS_QUOTE, IS_SPACE, IS_TITLE, \
+    IS_OOV, IS_UPPER, IS_STOP, IS_CURRENCY, IS_LEFT_PUNCT, IS_RIGHT_PUNCT, IS_SENT_START, LIKE_NUM, LIKE_EMAIL, \
+    LANG, NORM, PREFIX, SENTIMENT, STRING, SUFFIX, TEXT_WITH_WS, WHITESPACE, LIKE_URL, MATCHER_SUPPORTED_ATTRIBUTES, \
+    ENT_ID, ENT_IOB, ENT_KB_ID, HAS_VECTOR
 from PatternOmatic.settings.log import LOG
 
 
 #
 # Dynamic Grammar (Backus Naur Form) Generator
 #
-def dynagg(samples: [Doc]) -> dict:
+def dynamic_generator(samples: [Doc]) -> dict:
     """
     Dynamically generates a grammar in Backus Naur Form (BNF) notation representing the available Spacy NLP
     Linguistic Feature values of the given sample list of Doc instances
@@ -32,7 +37,7 @@ def dynagg(samples: [Doc]) -> dict:
     # Update times token per pattern [Min length of tokens, Max length of tokens] interval
     pattern_grammar[P] = _symbol_stacker(T, max_length_token, min_length_token)
 
-    #  Update times features per token (Max length of features)
+    # Update times features per token (Max length of features)
     pattern_grammar[T] = _symbol_stacker(F, _get_features_per_token(features_dict))
 
     if config.use_token_wildcard is True:
@@ -83,7 +88,7 @@ def dynagg(samples: [Doc]) -> dict:
 #
 # BNF Utilities
 #
-def _features_seen(samples: [Doc]) -> int and dict:
+def _features_seen(samples: [Doc]) -> (int, int, dict, dict):
     """
     Builds up a dictionary containing Spacy Linguistic Feature Keys and their respective seen values for the sample
     Args:
