+++ tests/test_individual.py
@@ -3,7 +3,7 @@ import unittest
 import spacy
 
 from PatternOmatic.ge.stats import Stats
-from PatternOmatic.nlp.engine import dynagg as dgg
+from PatternOmatic.nlp.bnf import dynamic_generator as dgg
 from PatternOmatic.ge.individual import Individual
 from PatternOmatic.settings.config import Config
 from PatternOmatic.settings.literals import FitnessType
